refactor: move is deeplinking to global app state
