+++ apps/ledger-live-mobile/src/actions/earn.ts
@@ -5,5 +5,3 @@ import type { EarnSetInfoModalPayload } from "./types";
 export const setEarnInfoModal = createAction<EarnSetInfoModalPayload>(
   EarnActionTypes.EARN_INFO_MODAL,
 );
-
-export const setIsDeepLinking = createAction<boolean>(EarnActionTypes.SET_IS_DEEP_LINKING);
