+++ apps/ledger-live-mobile/src/context/AuthPass/index.tsx
@@ -6,9 +6,9 @@ import { withTranslation } from "react-i18next";
 import { createStructuredSelector } from "reselect";
 import { compose } from "redux";
 import { privacySelector } from "../../reducers/settings";
-import { isDeepLinkingSelector } from "../../reducers/earn";
+import { isDeepLinkingSelector } from "../../reducers/appstate";
 import { SkipLockContext } from "../../components/behaviour/SkipLock";
-import type { Privacy, State as GlobalState, EarnState } from "../../reducers/types";
+import type { Privacy, State as GlobalState, AppState as EventState } from "../../reducers/types";
 import AuthScreen from "./AuthScreen";
 import RequestBiometricAuth from "../../components/RequestBiometricAuth";
 
@@ -16,7 +16,7 @@ const mapStateToProps = createStructuredSelector<
   GlobalState,
   {
     privacy: Privacy | null | undefined;
-    isDeepLinking: EarnState["isDeepLinking"]; // skips screen lock for internal deeplinks from ptx web player.
+    isDeepLinking: EventState["isDeepLinking"]; // skips screen lock for internal deeplinks from ptx web player.
   }
 >({
   privacy: privacySelector,
@@ -37,7 +37,7 @@ type OwnProps = {
 type Props = OwnProps & {
   t: TFunction;
   privacy: Privacy | null | undefined;
-  isDeepLinking: EarnState["isDeepLinking"];
+  isDeepLinking: EventState["isDeepLinking"];
 };
 // as we needs to be resilient to reboots (not showing unlock again after a reboot)
 // we need to store this global variable to know if we need to isLocked initially
