+++ apps/ledger-live-mobile/src/reducers/appstate.ts
@@ -7,13 +7,14 @@ import type { AppState, State } from "./types";
 import type {
   AppStateAddBackgroundEventPayload,
   AppStateIsConnectedPayload,
+  AppStateIsDeeplinkingPayload,
   AppStatePayload,
   AppStateSetHasConnectedDevicePayload,
   AppStateSetModalLockPayload,
   AppStateUpdateMainNavigatorVisibilityPayload,
   DangerouslyOverrideStatePayload,
 } from "../actions/types";
-import { AppStateActionTypes } from "../actions/types";
+import { AppStateActionTypes, EarnActionTypes } from "../actions/types";
 
 export type AsyncState = {
   isConnected: boolean | null;
@@ -26,6 +27,7 @@ export const INITIAL_STATE: AppState = {
   backgroundEvents: [],
   debugMenuVisible: false,
   isMainNavigatorVisible: true,
+  isDeepLinking: false,
 };
 
 const handlers: ReducerMap<AppState, AppStatePayload> = {
@@ -80,6 +82,17 @@ const handlers: ReducerMap<AppState, AppStatePayload> = {
     isMainNavigatorVisible: (action as Action<AppStateUpdateMainNavigatorVisibilityPayload>)
       .payload,
   }),
+
+  /** Prevents deep links from triggering privacy lock. */
+  [AppStateActionTypes.SET_IS_DEEP_LINKING]: (state, action) => ({
+    ...state,
+    isDeepLinking: (action as Action<AppStateIsDeeplinkingPayload>)?.payload || false,
+  }),
+
+  [EarnActionTypes.EARN_INFO_MODAL]: state => ({
+    ...state,
+    isDeepLinking: true,
+  }),
 };
 
 // Selectors
@@ -102,5 +115,7 @@ export const networkErrorSelector = createSelector(
   isConnectedSelector,
   (isConnected: boolean | null) => (!isConnected ? globalNetworkDown : null),
 );
+
+export const isDeepLinkingSelector = (state: State) => state.appstate.isDeepLinking;
 
 export default handleActions<AppState, AppStatePayload>(handlers, INITIAL_STATE);
