+++ apps/ledger-live-mobile/src/reducers/types.ts
@@ -69,6 +69,8 @@ export type AppState = {
   modalLock: boolean;
   backgroundEvents: Array<FwUpdateBackgroundEvent>;
   isMainNavigatorVisible: boolean;
+  /** For deep links that inadvertently trigger privacy lock. Reset to false on close. */
+  isDeepLinking: boolean;
 };
 
 // === BLE STATE ===
@@ -268,8 +270,6 @@ export type EarnState = {
     message?: string;
     messageTitle?: string;
   };
-  /** Earn web player deeplinks need to skip privacy lock. Resets on return to Earn dashboard. If other deeplinks also need to skip lock, could move this to AppState or somewhere more global. */
-  isDeepLinking: boolean;
 };
 
 // === PROTECT STATE ===
