+++ apps/ledger-live-mobile/src/screens/PTX/Earn/EarnInfoDrawer.tsx
@@ -6,17 +6,19 @@ import { Track } from "../../../analytics";
 import QueuedDrawer from "../../../components/QueuedDrawer";
 import { useDispatch, useSelector } from "react-redux";
 import { earnInfoModalSelector } from "../../../reducers/earn";
-import { setEarnInfoModal, setIsDeepLinking } from "../../../actions/earn";
+import { setEarnInfoModal } from "../../../actions/earn";
+import { setIsDeepLinking } from "../../../actions/appstate";
 
 export function EarnInfoDrawer() {
   const { t } = useTranslation();
   const dispatch = useDispatch();
   const [modalOpened, setModalOpened] = useState(false);
+
   const openModal = useCallback(() => setModalOpened(true), []);
+
   const closeModal = useCallback(async () => {
     await dispatch(setEarnInfoModal({}));
     await setModalOpened(false);
-
     dispatch(setIsDeepLinking(false));
   }, [dispatch]);
   const { message, messageTitle } = useSelector(earnInfoModalSelector);
